# typescript-sdk/openapi/radioso.yaml
openapi: 3.1.0
info:
  title: radioso API
  version: 0.1.0
  description: Code-generated OpenAPI contract for the radioso backend
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: radioso_session
    workspaceSelection:
      type: apiKey
      in: header
      name: X-Workspace-Id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
          required:
            - code
            - message
      required:
        - error
...
    AssistantSettingsSection:
      type: object
      properties:
        assistantName:
          type: string
        greetingInstruction:
          type: string
        assistantDefaultLocale:
          type:
            - string
            - "null"
        proactiveGreetingEnabled:
          type: boolean
        assistantBootstrapActive:
          type: boolean
          description: Server-managed bootstrap readiness derived from the current assistant configuration.
          readOnly: true
        suggestedQuestionsEnabled:
          type: boolean
        customInstruction:
          type: string
        assistantLogoUrl:
          type:
            - string
            - "null"
      required:
        - assistantName
        - greetingInstruction
        - assistantDefaultLocale
        - proactiveGreetingEnabled
        - assistantBootstrapActive
        - suggestedQuestionsEnabled
        - customInstruction
        - assistantLogoUrl
    PlatformRetrievalSettingsSection:
      type: object
      properties:
        queryRewriteEnabled:
          type: boolean
        semanticRewriteInstructions:
          type: string
          maxLength: 2000
        lexicalRewriteInstructions:
          type: string
          maxLength: 2000
        rerankEnabled:
          type: boolean
        vectorTopK:
          type: integer
          minimum: 1
          maximum: 300
        similarityThreshold:
          type: number
          minimum: 0
          maximum: 1
        rerankTopK:
          type: integer
          minimum: 1
        citationDisplayEnabled:
          type: boolean
        metadataRules:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              field:
                type: string
              valueType:
                type: string
                enum: *a1
              operator:
                type: string
                enum: *a2
              value:
                type: string
              combinator:
                type: string
                enum:
                  - and
                  - or
                default: and
              conditions:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    field:
                      type: string
                    valueType:
                      type: string
                      enum: *a1
                    operator:
                      type: string
                      enum: *a2
                    value:
                      type: string
                  required:
                    - id
                    - field
                    - valueType
                    - operator
                    - value
                default: []
              effect:
                type: string
                enum: *a3
              enabled:
                type: boolean
              triggerMode:
                type: string
                enum: *a4
              triggerInstruction:
                type: string
            required:
              - id
              - field
              - valueType
              - operator
              - value
              - effect
              - enabled
              - triggerMode
          default: []
        metadataFieldSuggestions:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              inferredType:
                type: string
                enum: *a1
            required:
              - field
              - inferredType
          default: []
      required:
        - queryRewriteEnabled
        - semanticRewriteInstructions
        - lexicalRewriteInstructions
        - rerankEnabled
        - vectorTopK
        - similarityThreshold
        - rerankTopK
        - citationDisplayEnabled
    PlatformChannelsSettingsSection:
      type: object
      properties:
        anonymousChatEnabled:
          type: boolean
        anonymousChatUrl:
          type:
            - string
            - "null"
        websiteEmbedEnabled:
          type: boolean
        websiteEmbedToken:
          type:
            - string
            - "null"
        websiteEmbedAllowedOrigins:
          type: array
          items:
            type: string
        websiteEmbedLauncherLabel:
          type: string
        websiteEmbedLauncherPosition:
          type: string
          enum: *a8
        websiteEmbedScriptUrl:
          type:
            - string
            - "null"
        websiteEmbedSnippet:
          type:
            - string
            - "null"
        websiteEmbedTheme:
          type: object
          properties:
            brand:
              type: string
            brandText:
              type: string
            surface:
              type: string
            text:
              type: string
          required:
            - brand
            - brandText
            - surface
            - text
        websiteEmbedCopy:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: string
        websiteEmbedExpertOverrides:
          type: object
          additionalProperties:
            type: string
      required:
        - anonymousChatEnabled
        - anonymousChatUrl
        - websiteEmbedEnabled
        - websiteEmbedToken
        - websiteEmbedAllowedOrigins
        - websiteEmbedLauncherLabel
        - websiteEmbedLauncherPosition
        - websiteEmbedScriptUrl
        - websiteEmbedSnippet
        - websiteEmbedTheme
        - websiteEmbedCopy
        - websiteEmbedExpertOverrides
    PlatformSettingsResponse:
      type: object
      properties:
        assistant:
          $ref: "#/components/schemas/AssistantSettingsSection"
        retrieval:
          $ref: "#/components/schemas/PlatformRetrievalSettingsSection"
        channels:
          $ref: "#/components/schemas/PlatformChannelsSettingsSection"
      required:
        - assistant
        - retrieval
        - channels
    UpdatePlatformSettingsRequest:
      type: object
      properties:
        assistant:
          type: object
          properties:
            assistantName:
              type: string
              maxLength: 200
            greetingInstruction:
              type: string
              maxLength: 200
            assistantDefaultLocale:
              type:
                - string
                - "null"
              maxLength: 35
            proactiveGreetingEnabled:
              type: boolean
            suggestedQuestionsEnabled:
              type: boolean
            customInstruction:
              type: string
              maxLength: 2000
        retrieval:
          type: object
          properties:
            queryRewriteEnabled:
              type: boolean
            semanticRewriteInstructions:
              type: string
              maxLength: 2000
            lexicalRewriteInstructions:
              type: string
              maxLength: 2000
            rerankEnabled:
              type: boolean
            vectorTopK:
              type: integer
            similarityThreshold:
              type: number
            rerankTopK:
              type: integer
            citationDisplayEnabled:
              type: boolean
            metadataRules:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                  field:
                    type: string
                    minLength: 1
                  valueType:
                    type: string
                    enum: *a1
                  operator:
                    type: string
                    enum: *a2
                  value:
                    type: string
                  combinator:
                    type: string
                    enum: *a5
                  conditions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          minLength: 1
                        field:
                          type: string
                          minLength: 1
                        valueType:
                          type: string
                          enum: *a1
                        operator:
                          type: string
                          enum: *a2
                        value:
                          type: string
                      required:
                        - id
                        - field
                        - valueType
                        - operator
                        - value
                  effect:
                    type: string
                    enum: *a3
                  enabled:
                    type: boolean
                  triggerMode:
                    type: string
                    enum: *a4
                  triggerInstruction:
                    type: string
                    maxLength: 500
                required:
                  - id
                  - effect
                  - enabled
        channels:
          type: object
          properties:
            anonymousChatEnabled:
              type: boolean
            websiteEmbedEnabled:
              type: boolean
            websiteEmbedAllowedOrigins:
              type: array
              items:
                type: string
                maxLength: 200
              maxItems: 20
            websiteEmbedLauncherLabel:
              type: string
              maxLength: 80
            websiteEmbedLauncherPosition:
              type: string
              enum: *a8
            websiteEmbedTheme:
              type: object
              properties:
                brand:
                  type: string
                  pattern: ^#[0-9a-fA-F]{6}$
                brandText:
                  type: string
                  pattern: ^#[0-9a-fA-F]{6}$
                surface:
                  type: string
                  pattern: ^#[0-9a-fA-F]{6}$
                text:
                  type: string
                  pattern: ^#[0-9a-fA-F]{6}$
            websiteEmbedCopy:
              type: object
              additionalProperties:
                type: object
                additionalProperties:
                  type: string
                  maxLength: 500
            websiteEmbedExpertOverrides:
              type: object
              additionalProperties:
                type: string
                maxLength: 500
    WorkspaceProviderCredentialSummary:
      type: object
      properties:
        provider:
          type: string
          enum: &a16
            - openai
            - openai-compatible
            - gemini
            - claude
        updatedAt:
          type: string
          format: date-time
      required:
        - provider
        - updatedAt
    WorkspaceProviderCredentialsResponse:
      type: object
      properties:
        encryptionConfigured:
          type: boolean
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceProviderCredentialSummary"
        envProviderAvailability:
          type: object
          properties:
            openai:
              type: boolean
            openai-compatible:
              type: boolean
            gemini:
              type: boolean
            claude:
              type: boolean
          required:
            - openai
            - openai-compatible
            - gemini
            - claude
      required:
        - encryptionConfigured
        - credentials
        - envProviderAvailability
    SetWorkspaceProviderCredentialRequest:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 1
          maxLength: 4096
      required:
        - apiKey
...
paths:
...
  /api/v1/settings:
    get:
      tags:
        - Settings
      summary: Get shared workspace platform settings
      operationId: getPlatformSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Shared assistant, retrieval, and channel settings returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Settings
      summary: Merge-update shared workspace platform settings
      operationId: updatePlatformSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlatformSettingsRequest"
      responses:
        "200":
          description: Shared settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/settings/credentials:
    get:
      tags:
        - Settings
      summary: List configured workspace provider API keys
      operationId: listWorkspaceProviderCredentials
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Configured providers and encryption status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceProviderCredentialsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/credentials/{provider}:
    put:
      tags:
        - Settings
      summary: Set or replace a workspace provider API key
      operationId: setWorkspaceProviderCredential
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum: *a16
          required: true
          name: provider
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetWorkspaceProviderCredentialRequest"
      responses:
        "204":
          description: Credential stored
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Server-side secret encryption is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Settings
      summary: Remove a stored workspace provider API key
      operationId: removeWorkspaceProviderCredential
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum: *a16
          required: true
          name: provider
          in: path
      responses:
        "204":
          description: Credential removed
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No credential found for the requested provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...